chore: sync fork (#313)
* chore: sync fork

* chore: sync fork

* chore: sync fork

* chore: sync fork

* chore: sync fork

* chore: sync fork

* chore: sync fork

* chore: sync fork

* chore: sync fork

+++ index.ts
@@ -4,7 +4,7 @@ import * as path from 'node:path'
 
 import { parseArgs } from 'node:util'
 import prompts from 'prompts'
-import { red, green, bold, yellow } from 'kolorist'
+import { red, green, bold, yellow } from 'kleur/colors'
 
 import ejs from 'ejs'
 
@@ -565,7 +565,7 @@ async function init() {
       root,
       () => {},
       (filepath) => {
-        if (filepath.endsWith('.js')) {
+        if (filepath.endsWith('.js') && !filepath.endsWith('eslint.config.js')) {
           const tsFilePath = filepath.replace(/\.js$/, '.ts')
           if (fs.existsSync(tsFilePath)) {
             fs.unlinkSync(filepath)
